backend/requirements.txt
asgiref==3.8.1
contourpy==1.3.2
csbdeep==0.8.1
cycler==0.12.1
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
fonttools==4.58.4
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
matplotlib==3.10.3
networkx==3.5
numba==0.61.2
numexpr==2.11.0
numpy==1.26.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stardist==0.9.1
tifffile==2025.6.11
tqdm==4.67.1
tzdata==2025.2
wheel==0.45.1
